refactor: 화장실 데이터를 서버 컴포넌트에서 fetching 하도록 변경
