+++ src/util/types/map.ts
@@ -1,0 +1,13 @@
+export interface Toilet {
+  POI_ID: string;
+  ANAME: string;
+  CENTER_X1: number;
+  CENTER_Y1: number;
+  CNAME: string;
+  FNAME: string;
+  INSERTDATE: string;
+  UPDATEDATE: string;
+  X_WGS84: number;
+  Y_WGS84: number;
+  DISTANCE: number;
+}
